rrr: convert misc/cpuinfo to module (rrr.cpuinfo)
cpuinfo.cpp is the rrr.cpuinfo module interface unit; cpuinfo.hpp
deleted. CPUInfo singleton reading /proc/<pid>/{net/dev,stat} and
/proc/meminfo for CPU / network / memory utilization. Uses
Log_debug from rrr.logging.

Sole consumer: src/rrr/rrr.hpp.

Measurement (rrr + rpcbench, -j32, clean):
  +alarm:    38.41 s, 175.5 MB PCM
  +cpuinfo:  39.60 s, 175.6 MB PCM

16 modules total; 2.35× baseline. Wallclock crept slightly (cpuinfo
has a large class body with rusty::Vec etc., bigger BMI than the
trivial earlier header-onlys).
