--- src/rrr/misc/cpuinfo.hpp
@@ -1,257 +1,0 @@
-#pragma once
-
-// import std; replacement — see <std_compat.hpp> for rationale.
-#include <std_compat.hpp>
-
-// @c-compat-added
-#include <cstddef>
-#include <cstdint>
-#include <cstdio>
-#include <cstdlib>
-#include <cstring>
-#include <ctime>
-
-#include <rusty/rusty.hpp>
-
-#include <sys/times.h>
-#include <unistd.h>
-
-#ifdef __linux__
-#include <sys/sysinfo.h>
-#endif
-
-
-
-
-#include "../base/all.hpp"
-
-namespace rrr {
-
-class CPUInfo {
-private:
-    unsigned long last_bytes_rxed[10], last_bytes_txed[10], last_mem_usage[10];
-    clock_t last_ticks_[10], last_user_ticks_[10], last_kernel_ticks_[10];
-		double last_cpu, last_txed, last_rxed, last_mem;
-		long long total_mem;
-		long page_size;
-		int index = 0;
-    pid_t pid_;
-		std::recursive_mutex mtx_;
-    //uint32_t num_processors_;
-    CPUInfo() {
-			const std::lock_guard<std::recursive_mutex> lock (mtx_);
-#ifdef __linux__
-    	struct tms tms_buf;
-			double txed, rxed;
-			rusty::Vec<double> result;
-			struct sysinfo mem_info;
-
-			sysinfo(&mem_info);
-			total_mem = mem_info.totalram;
-			total_mem *= mem_info.mem_unit;
-			total_mem /= 1024;
-			Log_debug("total amount of ram is: %lld", total_mem);
-
-			page_size = sysconf(_SC_PAGE_SIZE) / 1024;
-
-			last_ticks_[index] = times(&tms_buf);
-			last_kernel_ticks_[index] = tms_buf.tms_stime;
-			last_user_ticks_[index] = tms_buf.tms_utime;
-			
-			pid_ = ::getpid();
-			get_network(std::to_string(pid_), result, last_ticks_[index]);
-			get_memory(std::to_string(pid_), result, last_ticks_[index]);
-
-			index++;
-			//proc_cpuinfo = fopen("/proc/cpuinfo", "r");
-      //num_processors_ = 0;
-      //while (fgets(line, 1024, proc_cpuinfo) != NULL)
-      //    if (strncmp(line, "processor", 9) == 0)
-      //        num_processors_++;
-      //fclose(proc_cpuinfo);
-#else
-			last_cpu = last_txed = last_rxed = last_mem = 0.0;
-			total_mem = 0;
-			page_size = 0;
-			index = 0;
-			pid_ = ::getpid();
-#endif
-    }
-
-    /**
-     * get current cpu utilization for the calling process
-     * result is calculated as (sys_time + user_time) / total_time, since the
-     * last recent call to this function
-     *
-     * return:  upon success, this function will return the utilization roughly
-     *          between 0.0 to 1.0; otherwise, it will return a negative value
-     */
-    rusty::Vec<double> get_cpu_stat() {
-			const std::lock_guard<std::recursive_mutex> lock (mtx_);
-
-      struct tms tms_buf;
-      clock_t ticks;
-			rusty::Vec<double> result;
-			unsigned long txed, rxed;
-      double cpu_total, tx_total, rx_total;
-      clock_t last_ticks;
-			
-			ticks = times(&tms_buf);
-			if(index < 10) last_ticks = last_ticks_[index-1];
-			else last_ticks = last_ticks_[9];
-
-      Log_debug("ticks: %d -> %d", last_ticks, ticks);
-			//clock_t last_secs = last_ticks/CLOCKS_PER_SEC;
-			//clock_t secs = ticks/CLOCKS_PER_SEC;
-      //Log_info("sec: %d -> %d", last_secs, secs);
-			if (ticks <= last_ticks + 60/* || num_processors_ <= 0*/){
-				if(index < 10){
-					return {-1.0, -1.0, -1.0, -1.0};
-				} else{
-					return {last_cpu, last_txed, last_rxed, last_mem};
-				}
-			}
-			
-			if(index < 10){
-				last_kernel_ticks_[index] = tms_buf.tms_stime;
-				last_user_ticks_[index] = tms_buf.tms_utime;
-				last_ticks_[index] = ticks;
-				index++;
-			} else{
-				for(int i = 0; i < 9; i++){
-					last_kernel_ticks_[i] = last_kernel_ticks_[i+1];
-					last_user_ticks_[i] = last_user_ticks_[i+1];
-					last_ticks_[i] = last_ticks_[i+1];
-				}
-				last_kernel_ticks_[9] = tms_buf.tms_stime;
-				last_user_ticks_[9] = tms_buf.tms_utime;
-				last_ticks_[9] = ticks;
-			}
-			
-			if(index < 10){
-				cpu_total = -1.0;
-			} else{	
-      	cpu_total = (tms_buf.tms_stime - last_kernel_ticks_[8]) +
-        	    (tms_buf.tms_utime - last_user_ticks_[8]);
-      	cpu_total /= (ticks - last_ticks_[8]);
-			}
-			last_cpu = cpu_total;
-        //ret /= num_processors_;
-	
-			if(index < 10) result.push(-1.0);
-			else result.push(cpu_total);
-			
-			get_network(std::to_string(pid_), result, ticks);
-			get_memory(std::to_string(pid_), result, ticks);
-    	return result;
-    }
-
-		void get_network(const std::string& pid, rusty::Vec<double>& result, clock_t ticks){
-#ifndef __linux__
-			(void) pid;
-			(void) ticks;
-			result.push(-1.0);
-			result.push(-1.0);
-			return;
-#else
-			double tx_total = -1.0, rx_total = -1.0;
-			std::string line;
-			std::string temp;
-			unsigned long txed, rxed;
-			std::ifstream netfile("/proc/"+pid+"/net/dev");
-
-			for(int i = 0; i < 4; i++){
-				getline(netfile, line);
-			}
-			
-			int i = 1;
-			char* token = strtok(&line[0], " ");
-			while(token != NULL){
-				token = strtok(NULL, " ");
-				if(i == 1){
-					txed = strtoul(token, NULL, 0);
-				}
-				if(i == 9){
-					rxed = strtoul(token, NULL, 0);
-					break;
-				}
-				i++;
-			}
-			
-			if(index < 10) {
-				last_bytes_txed[index] = txed;
-				last_bytes_rxed[index] = rxed;
-			} else{
-				for(int i = 0; i < 9; i++){
-					last_bytes_txed[i] = last_bytes_txed[i+1];
-					last_bytes_rxed[i] = last_bytes_rxed[i+1];
-				}
-				last_bytes_txed[9] = txed;
-				last_bytes_rxed[9] = rxed;
-			}
-
-			if(ticks != last_ticks_[0]){
-				if(index < 10){
-					tx_total = -1.0;
-					rx_total = -1.0;
-				} else{
-					tx_total = (txed-last_bytes_txed[8])/(ticks - last_ticks_[8]);
-					rx_total = (rxed-last_bytes_rxed[8])/(ticks - last_ticks_[8]);
-				}
-			}
-			
-			result.push(tx_total);
-			result.push(rx_total);
-
-			last_txed = tx_total;
-			last_rxed = rx_total;
-#endif
-		}
-
-		void get_memory(const std::string& pid, rusty::Vec<double>& result, clock_t ticks){
-#ifndef __linux__
-			(void) pid;
-			(void) ticks;
-			result.push(-1.0);
-			return;
-#else
-			long rss;
-			double mem_usage, mem_total = -1.0;
-			std::string ignore;
-
-			std::ifstream stat_file("/proc/"+pid+"/stat", std::ios_base::in);
-			
-			stat_file >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore
-							  >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore
-								>> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> ignore >> rss;
-			
-			mem_usage = rss * page_size;
-			
-			if(index < 10) {
-				last_mem_usage[index] = mem_usage;
-			} else{
-				for(int i = 0; i < 9; i++){
-					last_mem_usage[i] = last_mem_usage[i+1];
-				}
-				last_mem_usage[9] = mem_usage;
-			}
-
-			if(ticks != last_ticks_[0]){
-				if(index < 10) mem_total = -1;
-				else mem_total = (mem_usage - last_mem_usage[8])/(ticks - last_ticks_[8]);
-			}
-			
-			result.push(mem_total);
-
-			last_mem = mem_total;
-#endif
-		}
-
-public:
-    static rusty::Vec<double> cpu_stat() {
-        static CPUInfo cpu_info;
-        return cpu_info.get_cpu_stat();
-    }
-};
-
-}
